160 adding jwt claims

+++ section-11/resources/item.py
@@ -1,13 +1,7 @@
-import os
-import sqlite3
-
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from flask_restful import Resource, reqparse
 
 from models.item import ItemModel
-
-dir_path = os.path.dirname(os.path.realpath(__file__))
-database_location = os.path.join(dir_path, 'data.db')
 
 
 class Item(Resource):
@@ -23,15 +17,19 @@ class Item(Resource):
                         required=True,
                         help='This field is required.')
 
-    @jwt_required()
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
+        
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
 
         return {'message': 'Item Deleted.'}
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -39,7 +37,7 @@ class Item(Resource):
 
         return {'message': 'Item not found'}, 404
 
-    @jwt_required()
+    @jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {
@@ -56,7 +54,7 @@ class Item(Resource):
 
         return item.json(), 201
 
-    @jwt_required()
+    @jwt_required
     def put(self, name):
         data = Item.parser.parse_args()
         item = ItemModel.find_by_name(name)
@@ -70,7 +68,7 @@ class Item(Resource):
 
 
 class Items(Resource):
-    @jwt_required()
+    @jwt_required
     def get(self):
         items = [item.json() for item in ItemModel.find_all()]
         # or
